uintmax_t, intmax_t 74 UCs left

+++ includes/libft.h
@@ -94,7 +94,7 @@ void				btree_apply_inorder(t_btree *root, void (*applyf)(int));
 void				btree_apply_postorder(t_btree *root, void (*applyf)(int));
 void				btree_apply_preorder(t_btree *root, void (*applyf)(int));
 void				btree_add_node(t_btree **root, t_btree *node);
-char				*ft_itoa_base(int value, int base);
-char				*ft_itoa_upper_base(int value, int base);
+char				*ft_itoa_base(uintmax_t value, int base);
+char				*ft_itoa_upper_base(uintmax_t value, int base);
 char				*ft_imaxtoa(intmax_t nb);
 #endif
